third_party/blink/web_tests/wpt_internal/prerender/restriction-geolocation.https.html

<!DOCTYPE html>
<title>Access to the Geolocation API is deferred</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('test-channel');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });
  const url = `resources/geolocation.https.html`;
  window.open(url, '_blank', 'noopener');

  const result = await gotMessage;
  const expected = [
    {
      event: 'started waiting navigator.geolocation.getCurrentPosition',
      prerendering: true
    },
    {
      event: 'prerendering change',
      prerendering: false
    },
    {
      event: 'getCurrentPosition error: [object GeolocationPositionError]',
      prerendering: false
    },
    {
      event: 'finished waiting navigator.geolocation.getCurrentPosition',
      prerendering: false
    },
  ];
  assert_equals(result.length, expected.length);
  for (let i = 0; i < result.length; i++) {
    assert_equals(result[i].event, expected[i].event, `event${i}`);
    assert_equals(result[i].prerendering, expected[i].prerendering,
      `prerendering${i}`);
  }
}, `the access to the Geolocation API should be deferred until the prerendered
   page is activated`);

</script>
</body>
